refactor(config): Add better config
