+++ lib/database.js
@@ -1,20 +1,21 @@
 const mongoose = require('mongoose');
 mongoose.Promise = require('bluebird');
 
-const config = require('./config');
+const config = require('../config');
 // Connecting MongoDB using mongoose to our application
-mongoose.connect(config.MONGO_CONNECTION_STRING);
+mongoose.connect(config.mongoConnectionString);
 
 // This callback will be triggered once the connection is successfully established to MongoDB
 mongoose.connection.on('connected', () => {
-  console.log(`Mongoose default connection open to ${config.MONGO_CONNECTION_STRING}`);
+  console.log(`Mongoose default connection open to ${config.mongoConnectionString}`);
 });
 
 mongoose.connection.on('error', (err) => {
   console.error({ err }, `Mongoose default connection error: ${err}`);
 });
 
-mongoose.connection.on('disconnected', () => console.log('Mongoose default connection disconnected'));
+mongoose.connection.on('disconnected', () =>
+  console.log('Mongoose default connection disconnected'));
 
 process.on('SIGINT', () => {
   mongoose.connection.close(() => {
