+++ lib/log.js
@@ -1,18 +1,16 @@
 const bunyan = require('bunyan');
 
-const config = require('./config');
+const { logging } = require('../config');
 
 const options = {
-  name: config.LOG_APP_NAME,
-  level: config.LOG_DISABLE ? (bunyan.FATAL + 1) : config.LOG_LEVEL, // turn off logging https://github.com/trentm/node-bunyan/pull/148#issuecomment-63018668
-  version: config.LOG_APP_VERSION,
+  name: logging.appName,
+  level: logging.disable ? bunyan.FATAL + 1 : logging.level, // turn off logging https://github.com/trentm/node-bunyan/pull/148#issuecomment-63018668
+  version: logging.appVersion,
   src: true,
-  obfuscate: [
-    'body.password'
-  ],
-  period: config.LOG_ROTATION_PERIOD,
-  count: config.LOG_COPIES_COUNT,
-  obfuscatePlaceholder: config.LOG_OBFUSCATE_PLACEHOLDER,
+  obfuscate: ['body.password'],
+  period: logging.rotationPeriod,
+  count: logging.copiesCount,
+  obfuscatePlaceholder: logging.obfuscatePlaceholder,
   streams: [],
   genReqId: req => req.id
 };
@@ -22,15 +20,15 @@ const stdout = { stream: process.stdout };
 if (process.env.NODE_ENV !== 'test') {
   options.streams.push(
     {
-      path: `${config.LOGS_PATH}/error-${config.LOG_APP_NAME}.log`,
+      path: `${logging.path}/error-${logging.appName}.log`,
       level: 'error'
     },
     {
-      path: `${config.LOGS_PATH}/info-${config.LOG_APP_NAME}.log`,
+      path: `${logging.path}/info-${logging.appName}.log`,
       level: 'info'
     },
     {
-      path: `${config.LOGS_PATH}/warn-${config.LOG_APP_NAME}.log`,
+      path: `${logging.path}/warn-${logging.appName}.log`,
       level: 'warn'
     }
   );
