+++ lib/main.js
@@ -1,7 +1,7 @@
 const http = require('http');
 
 const app = require('./app');
-const config = require('./config');
+const config = require('../config');
 
 const server = http.createServer(app);
